Move basket.rb, warehouse.rb, product.rb to model folder and update them
